+++ pl_Backend/pom.xml
@@ -61,14 +61,6 @@
                     </excludes>
                 </configuration>
             </plugin>
-
-            <!-- 컨테이너 빌드 -->
-            <plugin>
-                <groupId>com.google.cloud.tools</groupId>
-                <artifactId>jib-maven-plugin</artifactId>
-                <version>2.8.0</version>
-            </plugin>
-
         </plugins>
     </build>
