+++ pl_Backend/src/main/java/com/dadokdadok/myapp/dadoktest/InputDTO.java
@@ -1,0 +1,14 @@
+package com.dadokdadok.myapp.dadoktest;
+
+public class InputDTO {
+    private String input;
+
+    // Getter와 Setter
+    public String getInput() {
+        return input;
+    }
+
+    public void setInput(String input) {
+        this.input = input;
+    }
+}
