--- pl_Backend/src/main/java/com/dadokdadok/myapp/test/TestController.java
@@ -1,22 +1,0 @@
-package com.dadokdadok.myapp.test;
-
-
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@Slf4j
-@RestController
-@RequestMapping("/api")
-@RequiredArgsConstructor
-public class TestController {
-
-    @GetMapping("/test/{testId}")
-    public String testConfig() {
-        return "pipe-line-test";
-
-    }
-}
